Refactor code to use Fluent API for Entity Framework configurations
In this commit, the code has been refactored to use the Fluent API provided by Entity Framework for configuring entities. This has allowed us to remove Data Annotations from the entity classes and centralize the configuration logic in the DbContext class. This makes the code more maintainable, cleaner, and more effective in handling complex entity configurations.