Refactor add movie/show feature and create PosterDownloader
The commit separates the AddHandler operation into two separate handlers: AddMovieHandler and AddShowHandler. The operation should allow users to add a movie and a TV show separately to increase modularity and readability of the code. In addition, PosterDownloader class has been created to handle poster downloads. This allows images to be retrieved from the corresponding API and saved locally more uniformly and efficiently. The ImagePath was also moved and updated to accommodate these updates. This change includes updates to the corresponding test cases.